Move gbz80 struct to dedicated core header
